+++ database/migrations/2025_08_18_133610_create_carts_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('carts', function (Blueprint $table) {
-            $table->id('cart_id');
-            $table->string('session_id');
-            $table->unsignedBigInteger('item_id');
-            $table->string('item_name');
-            $table->decimal('item_price', 10, 2);
-            $table->integer('quantity');
-            $table->decimal('subtotal', 10, 2);
-            $table->timestamps();
-            
-            $table->foreign('item_id')->references('item_id')->on('master_items')->onDelete('cascade');
-            $table->index('session_id');
-        });
+        if (!Schema::hasTable('carts')) {
+            Schema::create('carts', function (Blueprint $table) {
+                $table->id('cart_id');
+                $table->string('session_id');
+                $table->unsignedBigInteger('item_id');
+                $table->string('item_name');
+                $table->decimal('item_price', 10, 2);
+                $table->integer('quantity');
+                $table->decimal('subtotal', 10, 2);
+                $table->timestamps();
+                
+                $table->foreign('item_id')->references('item_id')->on('master_items')->onDelete('cascade');
+                $table->index('session_id');
+            });
+        }
     }
 
     /**
